moving pipleines to separate folder and restoring EBOLA examples
