+++ example_data/20181008_1405_EBOV/run.sh
@@ -1,0 +1,6 @@
+#!/usr/bin/env bash
+
+echo "Running EBOV test set."
+echo "(Ensure RAMPART client is running in dev mode in another terminal)"
+
+rampart --devClient  --verbose --protocol ../../example_protocols/EBOV --clearAnnotated
