+++ example_protocols/EBOV/protocol.json
@@ -1,0 +1,5 @@
+{
+  "name": "ARTIC Ebola virus protocol v1.1",
+  "description": "Amplicon based sequencing of Ebola virus (Zaire species).",
+  "url": "http://artic.network/"
+}
