--- old/example_data/20181008_1405_EBOV/run_configuration.json
@@ -1,26 +1,0 @@
-{
-  "title": "EBOV Validation Run",
-  "basecalledPath": "fastq/pass",
-  "samples": [
-    {
-      "name": "Mayinga",
-      "description": "",
-      "barcodes": [ "NB01" ]
-    },
-    {
-      "name": "Kikwit",
-      "description": "",
-      "barcodes": [ "NB03" ]
-    },
-    {
-      "name": "Makona",
-      "description": "",
-      "barcodes": [ "NB04" ]
-    },
-    {
-      "name": "Negative Control",
-      "description": "",
-      "barcodes": [ "NB02" ]
-    }
-  ]
-}
