--- old/example_protocols/EBOV/primers.json
@@ -1,28 +1,0 @@
-{
-	"name": "EBOV primer scheme v1.0",
-	"amplicons": [
-		[32, 1057],
-		[907, 1881],
-		[1730, 2711],
-		[2555, 3578],
-		[3417, 4431],
-		[4274, 5301],
-		[5149, 6052],
-		[5895, 6814],
-		[6656, 7614],
-		[7438, 8428],
-		[8239, 9289],
-		[9139, 10048],
-		[9895, 10809],
-		[10651, 11668],
-		[11517, 12489],
-		[12337, 13298],
-		[13141, 14053],
-		[13897, 14803],
-		[14651, 15643],
-		[15480, 16493],
-		[16331, 17333],
-		[17182, 18183],
-		[17941, 18916]
-	]
-}
