+++ pipelines/default_pipeline/demux_map/config.yaml
@@ -1,0 +1,30 @@
+##### Fastq file stem #####
+
+filename_stem: sample_reads 
+
+##### Where is the data? #####
+
+input_path: path/to/data #where rampart will find the data
+output_path: path/to/data #where rampart will put the data
+
+##### Mapping reference panel #####
+
+references_file: references.fasta
+
+##### Porechop options #####
+
+require_two_barcodes: "True"
+discard_middle: "False"
+split_reads: "False"
+discard_unassigned: "False" #default keep unassigned, change to --discard_unassigned if you don't want to see unbarcoded reads
+barcode_set: "native" # [none,native,rapid,pcr,all], defaults to `native`, `all` is much slower
+limit_barcodes_to: #barcode csv string e.g. NB01,NB02,NB03,NB04,NB05,NB06
+barcode_threshold: 80
+barcode_diff: 5
+threads: 2
+
+
+##### Filtering options #####
+
+min_read_length: 0
+max_read_length: 1000000
